15h30 16/04 test be ok

+++ backend/src/main/java/com/fastfood/dto/request/FoodIngredientRequest.java
@@ -2,8 +2,10 @@ package com.fastfood.dto.request;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class FoodIngredientRequest {
     private String idIngredient;
-    private Double quantityUsed;
+    private BigDecimal quantityUsed;
 }
