+++ backend/src/main/java/com/fastfood/dto/request/IngredientRequest.java
@@ -4,11 +4,13 @@ import lombok.AccessLevel;
 import lombok.Data;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Data
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class IngredientRequest {
     String imageUrlIngredient;
     String ingredientName;
     String unit;
-    int quantityStock;
+    BigDecimal quantityStock;
 }
